Move reactor persistence models to individual files
