+++ st2actions/tests/unit/test_executions.py
@@ -28,7 +28,8 @@ from st2common.models.db.action import LiveActionDB
 from st2common.models.api.trigger import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
 from st2common.models.api.rule import RuleAPI
 from st2common.models.api.action import RunnerTypeAPI, ActionAPI
-from st2common.persistence.reactor import TriggerType, Trigger, TriggerInstance, Rule
+from st2common.persistence.rule import Rule
+from st2common.persistence.trigger import TriggerType, Trigger, TriggerInstance
 from st2common.persistence.action import RunnerType, Action, LiveAction
 from st2common.persistence.execution import ActionExecution
 from st2common.services import action as action_service
@@ -65,7 +66,7 @@ class TestActionExecutionHistoryWorker(DbTestCase):
         Action.add_or_update(ActionAPI.to_model(action_chain))
 
     def tearDown(self):
-        MOCK_FAIL_EXECUTION_CREATE = False    # noqa
+        MOCK_FAIL_EXECUTION_CREATE = False      # noqa
         super(TestActionExecutionHistoryWorker, self).tearDown()
 
     def test_basic_execution(self):
