+++ st2api/st2api/controllers/v1/triggers.py
@@ -22,7 +22,7 @@ from st2common import log as logging
 from st2common.models.api.trigger import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
 from st2common.models.api.base import jsexpose
 from st2common.models.system.common import ResourceReference
-from st2common.persistence.reactor import TriggerType, Trigger, TriggerInstance
+from st2common.persistence.trigger import TriggerType, Trigger, TriggerInstance
 from st2common.services import triggers as TriggerService
 from st2api.controllers import resource
 from st2common.exceptions.apivalidation import ValueValidationException
