+++ st2common/st2common/models/api/rule.py
@@ -22,7 +22,7 @@ from st2common.models.api.trigger import TriggerAPI
 from st2common.models.api.tag import TagsHelper
 from st2common.models.db.rule import RuleDB, ActionExecutionSpecDB
 from st2common.models.system.common import ResourceReference
-from st2common.persistence.reactor import Trigger
+from st2common.persistence.trigger import Trigger
 import st2common.services.triggers as TriggerService
 from st2common.util import reference
 import st2common.validators.api.reactor as validator
