+++ st2common/st2common/services/executions.py
@@ -35,7 +35,8 @@ import st2common.util.action_db as action_utils
 from st2common.constants.action import LIVEACTION_STATUS_CANCELED
 from st2common.persistence.execution import ActionExecution
 from st2common.persistence.action import RunnerType
-from st2common.persistence.reactor import TriggerType, Trigger, TriggerInstance, Rule
+from st2common.persistence.rule import Rule
+from st2common.persistence.trigger import TriggerType, Trigger, TriggerInstance
 from st2common.models.api.action import RunnerTypeAPI, ActionAPI, LiveActionAPI
 from st2common.models.api.rule import RuleAPI
 from st2common.models.api.trigger import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
