+++ st2reactor/st2reactor/bootstrap/sensorsregistrar.py
@@ -21,7 +21,7 @@ from st2common import log as logging
 from st2common.bootstrap.base import ResourceRegistrar
 import st2common.content.utils as content_utils
 from st2common.models.api.sensor import SensorTypeAPI
-from st2common.persistence.reactor import SensorType
+from st2common.persistence.sensor import SensorType
 
 __all__ = [
     'SensorsRegistrar',
