+++ st2reactor/st2reactor/cmd/sensormanager.py
@@ -9,7 +9,7 @@ from st2common.constants.logging import DEFAULT_LOGGING_CONF_PATH
 from st2common.exceptions.sensors import SensorNotFoundException
 from st2common.models.db import db_setup
 from st2common.models.db import db_teardown
-from st2common.persistence.reactor import SensorType
+from st2common.persistence.sensor import SensorType
 from st2common.signal_handlers import register_common_signal_handlers
 from st2reactor.sensor import config
 from st2common.transport.utils import register_exchanges
