+++ st2reactor/st2reactor/container/utils.py
@@ -16,7 +16,7 @@
 import six
 
 from st2common import log as logging
-from st2common.persistence.reactor import TriggerInstance
+from st2common.persistence.trigger import TriggerInstance
 from st2common.models.db.trigger import TriggerInstanceDB
 from st2common.services import triggers as TriggerService
