+++ st2reactor/tests/unit/test_rule_engine.py
@@ -20,7 +20,8 @@ from mongoengine import NotUniqueError
 
 from st2common.models.api.rule import RuleAPI
 from st2common.models.db.trigger import (TriggerDB, TriggerTypeDB)
-from st2common.persistence.reactor import (TriggerType, Trigger, Rule)
+from st2common.persistence.rule import Rule
+from st2common.persistence.trigger import (TriggerType, Trigger)
 import st2reactor.container.utils as container_utils
 from st2reactor.rules.enforcer import RuleEnforcer
 from st2reactor.rules.engine import RulesEngine
