+++ st2reactor/tests/unit/test_rule_matcher.py
@@ -17,7 +17,8 @@ import datetime
 
 from st2common.models.api.rule import RuleAPI
 from st2common.models.db.trigger import (TriggerDB, TriggerTypeDB)
-from st2common.persistence.reactor import (TriggerType, Trigger, Rule)
+from st2common.persistence.rule import Rule
+from st2common.persistence.trigger import (TriggerType, Trigger)
 from st2common.services.triggers import get_trigger_db_by_ref
 import st2reactor.container.utils as container_utils
 from st2reactor.rules.matcher import RulesMatcher
